2432: refactor TemplateMapper
Divides TemplateMapper more appropriately between service and mapper.
