+++ src/main/java/se/sundsvall/byggrintegrator/service/template/TemplateMapper.java
@@ -2,7 +2,6 @@ package se.sundsvall.byggrintegrator.service.template;
 
 import generated.se.sundsvall.arendeexport.v4.HandelseHandling;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
@@ -11,30 +10,15 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import org.springframework.stereotype.Component;
-import org.thymeleaf.ITemplateEngine;
-import org.thymeleaf.context.Context;
-import se.sundsvall.byggrintegrator.model.ByggrErrandDto;
 import se.sundsvall.byggrintegrator.model.FileTemplateDto;
-import se.sundsvall.byggrintegrator.service.FileAccessTokenService;
-
-import static java.util.Optional.ofNullable;
 
 @Component
 public class TemplateMapper {
 
-	// Name of the file Thymeleaf should use as a template
-	private static final String TEMPLATE_FILE = "neighborhood-notification-file-list";
-	private static final String DESCRIPTION_PROPERTY_DESIGNATION_TEMPLATE_FILE = "description-property-designation";
-	private static final Locale LOCALE = Locale.of("sv", "SE");
+	private final FileUrlService fileUrlService;
 
-	private final TemplateProperties templateProperties;
-	private final ITemplateEngine templateEngine;
-	private final FileAccessTokenService fileAccessTokenService;
-
-	public TemplateMapper(final TemplateProperties templateProperties, final ITemplateEngine templateEngine, final FileAccessTokenService fileAccessTokenService) {
-		this.templateProperties = templateProperties;
-		this.templateEngine = templateEngine;
-		this.fileAccessTokenService = fileAccessTokenService;
+	public TemplateMapper(FileUrlService fileUrlService) {
+		this.fileUrlService = fileUrlService;
 	}
 
 	private static <T> Predicate<T> distinctByKey(final Function<? super T, ?> keyExtractor) {
@@ -42,48 +26,8 @@ public class TemplateMapper {
 		return t -> seen.add(keyExtractor.apply(t));
 	}
 
-	/**
-	 * Generate a list of files as HTML
-	 *
-	 * @param  byggrErrandDto The errand to generate the file list for
-	 * @param  handlingtyper  The handling typer to use when generating the file list
-	 * @return                The HTML as a string
-	 */
-	public String generateFileList(final String municipalityId, final ByggrErrandDto byggrErrandDto, final Map<String, String> handlingtyper, final List<HandelseHandling> handling) {
-		final var fileTemplateDtoList = ofNullable(byggrErrandDto)
-			.map(errandDto -> mapByggrFilesToList(municipalityId, handlingtyper, handling))
-			.orElse(List.of());
-
-		final var context = new Context(LOCALE);
-		// Add the list of files to the context (as "fileList") so Thymeleaf can use it
-		context.setVariable("fileList", fileTemplateDtoList);
-		context.setVariable("supplementaryHeader", createSupplementaryHeader(byggrErrandDto));
-
-		return templateEngine.process(TEMPLATE_FILE, context);
-	}
-
-	public String getDescriptionAndPropertyDesignation(final ByggrErrandDto byggrErrandDto) {
-		final var context = new Context(LOCALE);
-		context.setVariable("descriptionAndPropertyDesignation", createSupplementaryHeader(byggrErrandDto));
-
-		return templateEngine.process(DESCRIPTION_PROPERTY_DESIGNATION_TEMPLATE_FILE, context);
-	}
-
-	/**
-	 * Create a supplementary header for the file list The supplementary header is the description of the errand
-	 * concatenated with the property designation
-	 *
-	 * @param  byggrErrandDto The errand
-	 * @return                The supplementary header
-	 */
-	private String createSupplementaryHeader(final ByggrErrandDto byggrErrandDto) {
-		return ofNullable(byggrErrandDto)
-			.map(header -> byggrErrandDto.getDescription() + " (" + byggrErrandDto.getPropertyDesignation() + ")")
-			.orElse("");
-	}
-
 	// Create a list that we iterate over with Thymeleaf
-	private List<FileTemplateDto> mapByggrFilesToList(final String municipalityId, final Map<String, String> handlingtyper, final List<HandelseHandling> handling) {
+	public List<FileTemplateDto> mapByggrFilesToList(final String municipalityId, final Map<String, String> handlingtyper, final List<HandelseHandling> handling) {
 		return handling.stream()
 			.map(entry -> mapToFileTemplateDto(municipalityId, parseDokId(entry), entry, handlingtyper))
 			.filter(distinctByKey(FileTemplateDto::getFileName)) // Remove duplicate file names if such exists
@@ -93,22 +37,9 @@ public class TemplateMapper {
 
 	private FileTemplateDto mapToFileTemplateDto(final String municipalityId, final int fileId, final HandelseHandling documentNameAndType, final Map<String, String> handlingtyper) {
 		return FileTemplateDto.builder()
-			.withFileUrl(parseFileUrl(municipalityId, fileId))
+			.withFileUrl(fileUrlService.parseFileUrl(municipalityId, fileId))
 			.withFileName(parseFileName(documentNameAndType, handlingtyper))
 			.build();
-	}
-
-	private String parseFileUrl(final String municipalityId, final int fileId) {
-		// String content is divided into the following: [domain][version]/[municipalityid][subdirectory][file
-		// identificator]?token=[uuid]
-		final var token = fileAccessTokenService.createToken(municipalityId, String.valueOf(fileId));
-		return "%s%s/%s%s%s?token=%s".formatted(
-			templateProperties.domain(),
-			templateProperties.version(),
-			municipalityId,
-			templateProperties.subDirectory(),
-			fileId,
-			token);
 	}
 
 	private String parseFileName(final HandelseHandling documentNameAndType, final Map<String, String> handlingtyper) {
